Remove dependency from DfConnect API -> DfWorkshop
