+++ Assets/Scripts/API/DFBlock.cs
@@ -10,9 +10,9 @@
 //
 
 #region Using Statements
-using DaggerfallWorkshop.Utility.AssetInjection;
 using FullSerializer;
 using System;
+using System.Collections.Generic;
 #endregion
 
 namespace DaggerfallConnect
@@ -1098,4 +1098,50 @@ namespace DaggerfallConnect
 
         #endregion
     }
+
+    #region FullSerializer Custom Processors (used when serializing world data structures)
+
+    public class RdbBlockDescProcessor : fsObjectProcessor
+    {
+        // Invoked after serialization has finished. Update any state inside of instance, modify the output data, etc.
+        public override void OnAfterSerialize(Type storageType, object instance, ref fsData data)
+        {
+            // Truncate ModelReferenceList at the first null entry in array[750].
+            fsData modelRefDict = data.AsDictionary["ModelReferenceList"];
+            if (!modelRefDict.IsNull)
+            {
+                List<fsData> modelRefList = modelRefDict.AsList;
+                for (int i = 0; i < modelRefList.Count; i++)
+                {
+                    if (modelRefList[i].AsDictionary["ModelIdNum"].AsInt64 == 0)
+                    {
+                        modelRefList.RemoveRange(i, modelRefList.Count - i);
+                        break;
+                    }
+                }
+            }
+        }
+    }
+
+    public class RdbObjectProcessor : fsObjectProcessor
+    {
+        // Invoked after serialization has finished. Update any state inside of instance, modify the output data, etc.
+        public override void OnAfterSerialize(Type storageType, object instance, ref fsData data)
+        {
+            // Only write relevant type resource data for Rdb Objects.
+            Dictionary<string, fsData> rdbObject = data.AsDictionary;
+            DFBlock.RdbResourceTypes type = (DFBlock.RdbResourceTypes)Enum.Parse(typeof(DFBlock.RdbResourceTypes), rdbObject["Type"].AsString);
+            Dictionary<string, fsData> resources = rdbObject["Resources"].AsDictionary;
+
+            if (type == DFBlock.RdbResourceTypes.Flat || type == DFBlock.RdbResourceTypes.Light)
+                resources.Remove("ModelResource");
+            if (type == DFBlock.RdbResourceTypes.Flat || type == DFBlock.RdbResourceTypes.Model)
+                resources.Remove("LightResource");
+            if (type == DFBlock.RdbResourceTypes.Model || type == DFBlock.RdbResourceTypes.Light)
+                resources.Remove("FlatResource");
+        }
+    }
+
+    #endregion
+
 }
